Validate DRPolicy against a single cluster
Update DRPolicy validation to require only one valid drCluster. This allows DRPolicy
to remain valid even if the other cluster are is unreachable. It supports scenarios
where the active hub and primary cluster go offline together, and the user chooses
to recover the hub. In such cases, the failed cluster may not be auto-imported, but
failover actions can still proceed.

Signed-off-by: Benamar Mekhissi <bmekhiss@ibm.com>